requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pgvector==0.3.4
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
PyMuPDF==1.24.10
python-docx==1.1.2
python-multipart==0.0.9
alembic==1.13.2
pytest==8.3.3
httpx==0.27.2